+++ public/js/app.js
@@ -6,13 +6,13 @@
 
 import SocketMethods   from './modules/app-socket.js?v=4.8.1';
 import UIBindMethods   from './modules/app-ui.js?v=4.8.1';
-import MediaMethods    from './modules/app-media.js?v=4.8.1';
+import MediaMethods    from './modules/app-media.js?v=4.8.2';
 import ContextMethods  from './modules/app-context.js?v=4.7.1';
 import ChannelMethods  from './modules/app-channels.js?v=4.8.1';
 import MessageMethods  from './modules/app-messages.js?v=4.8.1';
 import UserMethods     from './modules/app-users.js?v=4.7.1';
 import VoiceMethods    from './modules/app-voice.js?v=4.8.1';
-import UtilityMethods  from './modules/app-utilities.js?v=4.8.1';
+import UtilityMethods  from './modules/app-utilities.js?v=4.8.2';
 import AdminMethods    from './modules/app-admin.js?v=4.8.1';
 import PlatformMethods from './modules/app-platform.js?v=4.8.1';
 import SearchMethods   from './modules/app-search.js?v=3.49.0';